Add emojis to post reaction section
