+++ packages/lesswrong/components/votes/EAEmojisVoteOnPost.tsx
@@ -1,8 +1,29 @@
 import React from "react";
 import { registerComponent, Components } from "../../lib/vulcan-lib";
+import { useVote } from "./withVote";
 import type { PostVotingComponentProps } from "../../lib/voting/votingSystems";
 
-const styles = (_theme: ThemeType) => ({
+const styles = (theme: ThemeType) => ({
+  root: {
+    display: "flex",
+  },
+  divider: {
+    borderLeft: theme.palette.border.grey300,
+    margin: "6px 8px 8px 16px",
+  },
+  reacts: {
+    flexGrow: 1,
+    display: "flex",
+    alignItems: "center",
+    fontSize: 16,
+    fontWeight: 500,
+    fontFamily: theme.palette.fonts.sansSerifStack,
+    color: theme.palette.grey[600],
+    "& svg": {
+      width: 14,
+      height: 14,
+    },
+  },
 });
 
 interface EAEmojisVoteOnPostProps extends PostVotingComponentProps {
@@ -14,14 +35,22 @@ const EAEmojisVoteOnPost = ({
   votingSystem,
   classes,
 }: EAEmojisVoteOnPostProps) => {
-  const {PostsVoteDefault} = Components;
+  const voteProps = useVote(document, "Posts", votingSystem);
+  const {PostsVoteDefault, EAReactsSection} = Components;
   return (
-    <div>
+    <div className={classes.root}>
       <PostsVoteDefault
         post={document}
+        votingSystem={votingSystem}
         useHorizontalLayout
       />
-      <div>Custom</div>
+      <div className={classes.divider} />
+      <div className={classes.reacts}>
+        <EAReactsSection
+          document={document}
+          voteProps={voteProps}
+        />
+      </div>
     </div>
   );
 }
